parseFloat

+++ app/containers/Transactions.js
@@ -193,9 +193,9 @@ export default class Transactions extends Component {
         }
 
           const newBalance = type == "expenditure" 
-                ? doc.data().total - parseInt(amount)
+                ? doc.data().total - parseFloat(amount)
                 :  type == "income"
-                      ? doc.data().total + parseInt(amount)
+                      ? doc.data().total + parseFloat(amount)
                       : doc.data().total;
 
         transaction.update(this.balance, {
@@ -216,7 +216,7 @@ export default class Transactions extends Component {
           return 0;
         }
 
-          const newAmount = doc.data().currAmount + parseInt(amount);
+          const newAmount = doc.data().currAmount + parseFloat(amount);
 
         transaction.update(this.budget.doc(docId), {
           currAmount: newAmount,
